weights are now beteween 0 and 1 + corrected error (if get_direction returns illegal value it works)

+++ Game_neural.h
@@ -10,7 +10,7 @@ class Game_neural : public Game
 
 private:
 
-  int compute_fitness();
+  float compute_fitness();
   int get_direction(int depth);
   int get_direction_1();
   int get_direction_2();
